backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
websockets==13.1
python-dotenv==1.0.1
supabase==2.3.4
PyPDF2==3.0.1
pymupdf==1.24.0
tiktoken==0.7.0
openai>=2.15.0
numpy>=1.24.0


